Refactored to improve ergonomics, and keep related traits closer together.
There is now a set of traits for `Functor*` that are more ergonomic to use. These have been givin the name `Functor*`
while the rest have been renamed to `Functor*Impl`. Despite the increase in ergonomics, the need to have
`F::Impl: FunctorImpl` is still required, but it should by implied by the `F: Functor<_>` constraint.

Added a simple macro to help implement the higher kinded type traits, doesn't work for everything though.